requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
groq
python-multipart
rank_bm25
langchain-classic
chromadb